Feat: Beginning of built in functions, creation of tellaPutChar.

+++ myLibc/myLibc.c
@@ -1,5 +1,5 @@
-int myStrlen(char *string){
-    int count = 0;
+size_t myStrlen(char *string){
+    size_t count = 0;
     while(string[count] != '\0'){
         count++;
     }
